third_party/blink/web_tests/external/wpt/html/rendering/widgets/input-checkbox-switch.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] appearance:none should work: display gets its initial value
  assert_equals: expected "inline" but got "inline-block"
Harness: the test ran to completion.
